Fixed testintegrationtime.py
